improvements to contribution guide
extracted guides into separate articles
changes in ordering
readme links to playbook contribution guide
document status
anchor on improve-the-playbook standard footer
introduction para on playbook
article on GH issues
REVIEW badge takes account of requiresReview flag
