+++ lib/_components/appCard/template.njk
@@ -3,21 +3,19 @@
 {% if params.items.length > 0 %}
 	<ul class="nhsuk-grid-row nhsuk-card-group{%- if params.classes %} {{ params.classes }}{% endif %}">
 		{% for item in params.items | sortByOrder(params.order) %}
+		{% if not item.data.status or item.data.status != 'DRAFT' %}
 		<li class="nhsuk-grid-column-one-third nhsuk-card-group__item">
 			<div class="nhsuk-card nhsuk-card--clickable">
 				<div class="nhsuk-card__content">
 					<div class="nhsuk-card__header">
 						<h2 class="nhsuk-card__heading nhsuk-heading-m">
-							<a class="nhsuk-card__link" href="{{ item.url | url | pretty }}">
-								{{ item.data.title }}
-							</a>
+							<a class="nhsuk-card__link" href="{{ item.url | url | pretty }}">{{ item.data.title }}</a>
 						</h2>
 						{{ appBadges({
 								badges: item.data.badges
 							}) 
 						}}
 					</div>
-					
 					{% if item.data.description %}
 					<p class="nhsuk-card__description">{{ item.data.description | markdown("inline") | safe }}</p>
 					{% endif %}
@@ -27,6 +25,7 @@
 				</div>
 			</div>
 		</li>
+		{% endif %}
 		{% endfor %}
 	</ul>
 {% else %}
