+++ lib/_computed/badges.js
@@ -1,3 +1,4 @@
+const { DateTime } = require('luxon')
 
 
 "nhsuk-tag--grey"
@@ -5,15 +6,25 @@ const styleLookup = {
     DRAFT: "nhsuk-tag--grey",
     REVIEW: "nhsuk-tag--yellow"
 }
-
 module.exports = function(data) {
-    result = []
-    if (data.status) {
-        let style = styleLookup[data.status];
-        result.push({
-            text: data.status,
-            style: style
-        })
+
+    function requiresReview(review, title) {
+        if (review &&
+            review.requires_review) {
+                return true;
+            }
     }
-    return result
+
+    var status;
+    if (requiresReview(data.review, data.title)) {
+        status = "REVIEW";
+    } else if ("FINAL" != data.status) {
+        status = data.status;
+    }
+    if (status) {
+        return [{
+            text: status,
+            style: styleLookup[status]
+        }]
+    }
 };
