+++ src/design/index.md
@@ -3,7 +3,7 @@ layout: article
 title: "Design"
 description: "How we design at the NHSBSA"
 tags: home
-order: 20
+order: 10
 related:
   title: Articles
   tag: design
