+++ src/development/coding-securely.md
@@ -7,6 +7,7 @@ order:
   dev: 2
   security: 1
 related:
+  title: Related articles
   tag: dev-security
 ---
 ## NCSC guidance
