+++ src/development/index.md
@@ -3,9 +3,8 @@ layout: article
 title: "Development"
 description: "How we develop at the NHSBSA"
 tags: home
-order: 20
+order: 30
 related:
-  title: Articles
   tag: dev
 ---
