+++ src/how-to-contribute/github-issues.md
@@ -1,0 +1,44 @@
+---
+layout: article
+title: "Suggest a change"
+description: "Use Github issues to suggest and manage change"
+tags: [playbook]
+order: 1
+---
+## Github Issues
+
+We use [Github issues][github_issues_nhsbsa_digital_playbook] to keep track of changes to the playbook. If you want to make a change, raise an issue and open it up for discussion.
+
+When raising an issue, consider:
+
+* __What__
+  Give a brief description of the proposal.
+
+* __Why__
+  Explain why you think this should be added to the Playbook
+  
+  * What evidence do you have that it's needed?
+  * What evidence do you have that it meets the needs of our people and the services they build?
+  * Have you checked that it doesn't already exist in the Playbook?
+
+* __Anything else__
+  Include links to any examples, research or code to support your proposal, if available.
+
+Every page in the playbook has information on how to [improve the playbook](#improve-the-playbook), with a link to create a Github issue.
+
+## Seek peer review
+
+Use NHSBSA Slack or Teams to draw attention to your suggestion and get feedback:
+
+* Is there really a need?
+* Is the proposal the right approach?
+* Can anyone help?
+
+## Draft content
+
+You can write some content and submit a pull request to get it published to the playbook website.
+
+By setting the article `status` as `DRAFT`, it will not appear in a referenced article and can only be viewed with a direct link.
+Articles at `REVIEW` status will appear as a referenced article but clearly marked with it's status.
+
+[github_issues_nhsbsa_digital_playbook]: <https://github.com/nhsbsa/nhsbsa-digital-playbook/issues>
