+++ src/how-to-contribute/information-architecture.md
@@ -1,0 +1,45 @@
+---
+layout: article
+title: "Information architecture"
+description: "Organise the playbook content into sections"
+tags: playbook
+order: 2
+---
+We have three levels in our Information Architecture:
+
+* Home
+* Section
+* Additional detail
+
+## Home
+
+The home page displays a list of articles that are jumping off points into different sections. These include
+
+* _Role based_
+  * Delivery (including agile)
+  * Content Design
+  * [Research](../../research)
+  * [Design](../../design)
+  * Architecture
+  * [Development](../../development)
+  * [Test](../../test)
+  * Platform
+  * Support
+* _People and ways of working_
+  * Communities
+  * Our people
+  * Admin
+  * Meetings
+* _Contributing_
+
+The Playbook is a living set of documents. Not all sections have been written.
+
+## Section
+
+Each section will contain articles relating to that topic.
+
+## Additional detail
+
+For more extensive and detailed content we may link from section articles into further specialist articles.
+
+For instance the _Developer/testing_ section article links to articles about _writing clean tests_ and _writing integration tests_.
