+++ src/research/index.md
@@ -4,8 +4,9 @@ title: "Research"
 description: "How we do research at the NHSBSA"
 tags: home
 order: 20
-related:
-  title: Articles
-  tag: research
 ---
+We have a separate playbook that is written and maintained by our user research community. In time, this will become one with this overall playbook.
 
+But for now, the link to this is here: [NHSBSA User Research Playbook][ur_playbook]
+
+[ur_playbook]: <https://nhsbsa-playbook.netlify.app/>
