--- src/research/ur-playbook.md
@@ -1,14 +1,0 @@
----
-layout: article
-title: "Research playbook"
-description: "What do we do when researching..."
-tags: research
-order: 1
----
-
-We already have a separate playbook that is written and maintained by our user research community. In time, this will become one with this overall playbook.
-
-But for now, the link to this is here: [NHSBSA User Research Playbook][ur_playbook]
-
-
-[ur_playbook]: <https://nhsbsa-playbook.netlify.app/>
