+++ src/standards/index.md
@@ -1,0 +1,10 @@
+---
+layout: article
+title: "Standards"
+description: "All our standards in one place"
+status: DRAFT
+tags: home
+order: 200
+related:
+  tag: standard
+---
