+++ src/testing/index.md
@@ -3,7 +3,7 @@ layout: collection
 title: "Testing"
 description: "How we test at the NHSBSA"
 tags: home
-order: 30
+order: 40
 collection_tag: test
 pagination:
   data: collections.test
